requirements.txt
openpyxl
pandas
numpy
beautifulsoup4
selenium
pyaml